Make the Updates class object oriented
Umm, I'm not entirely sure I understand this, but for some reason, when
I was making updates and then syncing them over to the webserver, it
wasn't updating the site unless I re-ran hypnotoad. And that was
annoying. Now this will initialize the DB handle every time you hit the
/updates page, which I guess is fine.