Hide certain translations from Flashcards
Mainly the ones where I just repeat the word. I don't know, I felt like
the flashcards were getting a little cluttered. But honestly, the code
for all this was confusing me so much that I'm not sure if this was
worth it.

I realized that get_all_translations_by_id is never called with a list
of words so I simplified that.

I also added a few comments...I'll address them later