Display 'notes' next to word if it's the only entry on the page
I was looking at the word "fly" and thought it was odd that you could
just see "insect" below the chinese and peng'im. I think it makes more
sense to be right next to the word "fly". Having the notes does make
sense if there are multiple entries though, like for the word "aunt".