Add a Translation table
yeah, it's the start of some crazy restructuring the database time

my thought for this is to have a table that links Teochew and English
entries, so it can be a many-to-many relationship. and then alternate
Chinese characters would only be linked to the Teochew table.

so for something like 内 that is used for both "home" and "inside",
I only have to include the alternate characters once