Simplify the english pages for 'inside' and 'outside'
I'm realizing that there are some translations that only really apply
to partial words. Using lai6 on its own will always mean "home", so it
doesn't make sense to include it in the page for "inside". It _can_
mean inside as one part of a larger word, but not on its own. For these
"partial word translations", I'm having them link to a different
English entry, with hidden = 1

This also means I can remove the extra notes I typed up for "inside"
and "outside"