Add translation for "to wear"
I was randomly going through some flashcards, and it showed me the word
"jacket" and I thought, I should add "coat" as a synonym. I also added
the translation for "to wear" because I thought of it and realized I
didn't have it.

It took me a little bit to figure out what exactly I was doing with
Audacity to create a recording so I'm starting out slow and just adding
a single translation here