Pick the adjective form of "white" on the Tones lesson page
I noticed it was displayed "beh sek" here in the top table, and that's
not what I want. Everything should just be one syllable because we're
showing the original word without tone change

Also, this was kind of hack-y. Maybe all that logic should be under lib/
rather than a template file