Hardcode 'I' and 'you' having no tone change
So we don't have to store this column in the database. It just seems
silly to when only two words have this attribute.

Honestly, I should just modify the sentence translations in the
database to explicitly include '|' after each 'I' and 'you', but I don't
feel like doing that right now