Add a "standard_pengim" column to the Chinese table
Since I speak with the gekion accent, I'm making it so I store the
gekion pronunciation by default, rather than having this hack-y thing
where I switch over to an "alternate" pronunciation.

If there's a word that does have a different non-gekion pronunciation, I
will note that in this new optional column, standard_pengim

Yeah, I feel like I should have given this a better name, but this is
what I came up with. This will have the non-gekion, or teochew accent
way of saying something if the gekion one is different