Reorganize logic for 'chinese' controller action
There was a subtle bug in that $details was getting auto-vivified to an
arrayref even if it was originally set to undef. I decided to rename the
variable to $chinese because it fits better with the name of the stashed
value. Now this correctly stashes an undef value for 'chinese' if we
don't have the character in the database