Change where we numify a search input
So I just added "the first time" and when I tried to initially search
it, it automatically changed to "the 1 time" which was awkward. So let's
numify only to check for an exact match, and only if there wasn't
already a match without numifying. If there is no match at all, we
search with the same term that the user inputted