Be able to get flashcards for a specific category
According to my old POD that I had written, I had meant to make this
possible a while ago. Maybe it was, but it broke. This change makes it
so you can navigate directly to this feature using the url, but I still
need to add links somewhere in the UI