Add bootstrap5 css and js files
Note that the css file is a custom file that I generated using sass. I
feel like I should be including the custom scss file that I wrote up for
this, but I only learned the basics of scss for this, and I honestly
don't know the best way to include scss for this project. So, here's the
compiled version.

I also noticed that there are a lot of css and js files in this project
that could be cleaned up. Another day.