Change navbar background color and make it full width
I had to reorganize this a little and move the navbar out of the main
container-md div that's in the wrapper template. Now there are three
separate container-md divs: 1) the navbar, 2) the search/translate
inputs, and 3) the main body of the page

I added a header element _outside_ of the navbar's container-md div and
set the background color for this. This makes it so the navbar appears
to be the full width of the page, but the text is still limited to the
width of the container