Add a site settings dropdown to the navbar menu
which doesn't actually do anything yet. That's in a future commit

This adds a little gear icon that you can click to show a dropdown,
which contains the UI for two different settings: light/dark mode and
traditional/simplified characters, each controlled by radio inputs

There are actually two separate buttons here. For larger screens, the
button is all the way to the right, away from the rest of the navbar
links. For smaller screens, it's right next to the search and hamburger
icons

I couldn't find a way to have both buttons display the same dropdown
html element, so the dropdown menu contents get duplicated in the
rendered html